feat: added docker
